Fixed a bug where max zoom calculation was being truncated

+++ core/src/com/github/eoinf/jiggen/webapp/screens/views/PuzzleView.java
@@ -128,8 +128,8 @@ public class PuzzleView implements ScreenView {
         puzzleViewModel.getWorldBoundsObservable().subscribe(new Consumer<GridPoint2>() {
             @Override
             public void accept(GridPoint2 worldBounds) {
-                float maxZoomX = worldBounds.x / viewport.getScreenWidth();
-                float maxZoomY = worldBounds.y / viewport.getScreenHeight();
+                float maxZoomX = worldBounds.x / (float)viewport.getScreenWidth();
+                float maxZoomY = worldBounds.y / (float)viewport.getScreenHeight();
 
                 float maxZoom = Math.max(maxZoomX, maxZoomY);
                 camera.setCameraBounds(worldBounds.x, worldBounds.y, maxZoom);
